# docs/api/control.openapi.json
{
  "info": {
    "title": "one-fetch Control API",
    "version": "0.1.0",
    "description": "Canonical one-fetch Control API. Runtime capabilities remain authoritative."
  },
  "openapi": "3.1.0",
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque administrator access token"
      },
      "executionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque execution token scoped to its own report"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/audit": {
      "get": {
        "security": [{ "adminBearer": [] }],
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 1, "maxLength": 1024 },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Audit event page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schemaVersion": { "type": "number", "enum": [1] },
                    "events": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "type": "object",
                            "properties": {
                              "schemaVersion": {
                                "type": "number",
                                "enum": [1]
                              },
                              "eventId": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 128
                              },
                              "occurredAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "recordedAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "category": {
                                "type": "string",
                                "enum": [
                                  "execution",
                                  "auth",
                                  "account",
                                  "config",
                                  "security",
                                  "system",
                                  "audit"
                                ]
                              },
                              "action": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 128
                              },
                              "outcome": {
                                "type": "string",
                                "enum": [
                                  "success",
                                  "denied",
                                  "failure",
                                  "partial",
                                  "unknown"
                                ]
                              },
                              "severity": {
                                "type": "string",
                                "enum": ["info", "warning", "error", "critical"]
                              },
                              "actor": {
                                "type": "object",
                                "properties": {
                                  "type": {
                                    "type": "string",
                                    "enum": [
                                      "admin",
                                      "execution-token",
                                      "system",
                                      "anonymous"
                                    ]
                                  },
                                  "actorId": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 128
                                  },
                                  "credentialId": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 128
                                  },
                                  "sessionFingerprint": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 256
                                  }
                                },
                                "required": ["type"],
                                "additionalProperties": false
                              },
                              "correlation": {
                                "type": "object",
                                "properties": {
                                  "requestId": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 128,
                                    "pattern": "^[A-Za-z0-9._:-]+$"
                                  },
                                  "reportId": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 128
                                  },
                                  "connectionId": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 128
                                  },
                                  "configVersion": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 256
                                  },
                                  "providerRequestId": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 256
                                  }
                                },
                                "additionalProperties": false
                              },
                              "request": {
                                "type": "object",
                                "properties": {
                                  "transport": {
                                    "type": "string",
                                    "enum": ["http", "websocket", "tcp", "tls"]
                                  },
                                  "method": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 32
                                  },
                                  "origin": {
                                    "type": "string",
                                    "maxLength": 2048
                                  },
                                  "path": {
                                    "type": "string",
                                    "maxLength": 8192
                                  },
                                  "query": {
                                    "type": "array",
                                    "items": {
                                      "type": "array",
                                      "prefixItems": [
                                        { "type": "string", "maxLength": 1024 },
                                        { "type": "string", "maxLength": 4096 }
                                      ]
                                    },
                                    "maxItems": 256
                                  },
                                  "headers": {
                                    "type": "array",
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "name": {
                                          "type": "string",
                                          "minLength": 1,
                                          "maxLength": 256
                                        },
                                        "value": {
                                          "type": "string",
                                          "maxLength": 2048
                                        }
                                      },
                                      "required": ["name", "value"],
                                      "additionalProperties": false
                                    },
                                    "maxItems": 256
                                  },
                                  "contentType": {
                                    "type": "string",
                                    "maxLength": 1024
                                  }
                                },
                                "required": ["transport"],
                                "additionalProperties": false
                              },
                              "decision": {
                                "type": "object",
                                "properties": {
                                  "decision": {
                                    "type": "string",
                                    "enum": ["allow", "deny"]
                                  },
                                  "source": {
                                    "type": "string",
                                    "enum": [
                                      "default",
                                      "system-rule",
                                      "user-rule"
                                    ]
                                  },
                                  "ruleId": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 128
                                  },
                                  "warnings": {
                                    "type": "array",
                                    "items": {
                                      "type": "string",
                                      "maxLength": 1024
                                    },
                                    "maxItems": 64
                                  }
                                },
                                "required": ["decision", "source", "warnings"],
                                "additionalProperties": false
                              },
                              "result": {
                                "type": "object",
                                "properties": {
                                  "source": {
                                    "type": "string",
                                    "enum": [
                                      "target",
                                      "relay",
                                      "vendor",
                                      "unknown"
                                    ]
                                  },
                                  "status": {
                                    "type": "integer",
                                    "minimum": 100,
                                    "maximum": 599
                                  },
                                  "stage": {
                                    "type": "string",
                                    "maxLength": 128
                                  },
                                  "code": { "type": "string", "maxLength": 128 }
                                },
                                "required": ["source"],
                                "additionalProperties": false
                              },
                              "change": {
                                "type": "object",
                                "properties": {
                                  "beforeVersion": {
                                    "type": "string",
                                    "maxLength": 256
                                  },
                                  "afterVersion": {
                                    "type": "string",
                                    "maxLength": 256
                                  },
                                  "beforeHash": {
                                    "type": "string",
                                    "pattern": "^[a-f0-9]{64}$"
                                  },
                                  "afterHash": {
                                    "type": "string",
                                    "pattern": "^[a-f0-9]{64}$"
                                  },
                                  "changedFields": {
                                    "type": "array",
                                    "items": {
                                      "type": "string",
                                      "minLength": 1,
                                      "maxLength": 256
                                    },
                                    "maxItems": 256
                                  }
                                },
                                "required": ["changedFields"],
                                "additionalProperties": false
                              },
                              "metrics": {
                                "type": "object",
                                "properties": {
                                  "requestBytes": {
                                    "type": "integer",
                                    "minimum": 0
                                  },
                                  "responseBytes": {
                                    "type": "integer",
                                    "minimum": 0
                                  },
                                  "redirects": {
                                    "type": "integer",
                                    "minimum": 0
                                  },
                                  "durationMs": {
                                    "type": "number",
                                    "minimum": 0
                                  }
                                },
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "schemaVersion",
                              "eventId",
                              "occurredAt",
                              "recordedAt",
                              "category",
                              "action",
                              "outcome",
                              "severity",
                              "actor",
                              "correlation"
                            ],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "integrity": {
                                "type": "object",
                                "properties": {
                                  "payloadHash": {
                                    "type": "string",
                                    "pattern": "^[a-f0-9]{64}$"
                                  },
                                  "signature": {
                                    "type": "string",
                                    "minLength": 1,
                                    "pattern": "^[A-Za-z0-9_-]+$"
                                  },
                                  "keyId": {
                                    "type": "string",
                                    "minLength": 1,
                                    "maxLength": 128
                                  }
                                },
                                "required": [
                                  "payloadHash",
                                  "signature",
                                  "keyId"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": ["integrity"],
                            "additionalProperties": false
                          }
                        ]
                      },
                      "maxItems": 500
                    },
                    "nextCursor": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 1024
                    }
                  },
                  "required": ["schemaVersion", "events"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid cursor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alerts": {
      "get": {
        "security": [{ "adminBearer": [] }],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "501": {
            "description": "Feature unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backups": {
      "get": {
        "security": [{ "adminBearer": [] }],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "501": {
            "description": "Feature unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "servers": [
    { "url": ".", "description": "The configured Control service base URL" }
  ],
...
}